+++ internal/provider/resource_iosxe_interface_ethernet_test.go
@@ -25,6 +25,8 @@ func TestAccIosxeInterfaceEthernet(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.address", "10.10.10.10"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.global", "false"),
 					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "helper_addresses.0.vrf", "VRF1"),
+					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "source_template.0.template_name", "TEMP1"),
+					resource.TestCheckResourceAttr("iosxe_interface_ethernet.test", "source_template.0.merge", "false"),
 				),
 			},
 			{
@@ -80,6 +82,10 @@ func testAccIosxeInterfaceEthernetConfig_all() string {
 		address = "10.10.10.10"
 		global = false
 		vrf = "VRF1"
+		}]
+		source_template = [{
+		template_name = "TEMP1"
+		merge = false
 		}]
   		depends_on = [iosxe_restconf.PreReq0, iosxe_restconf.PreReq1, ]
 	}